pipecat-agent/pyproject.toml
[project]
name = "human-virtual-agent"
version = "0.1.0"
description = "Pipecat voice agent for Human Virtual avatars"
requires-python = ">=3.12"
dependencies = [
    "pipecat-ai[deepgram,openai,cartesia,silero]",
    "pipecat-ai-small-webrtc-prebuilt",
    "httpx",
    "python-dotenv",
    "loguru",
]
